datapath: Convert ODP_VPORT_* to use AF_NETLINK socket layer.
This commit calls genl_lock() and thus doesn't support Linux before
2.6.35, which wasn't exported before that version.  That problem will
be fixed once the whole userspace interface transitions to Generic
Netlink a few commits from now.

Acked-by: Jesse Gross <jesse@nicira.com>

+++ lib/dpif-linux.h
@@ -18,13 +18,14 @@
 #define DPIF_LINUX_H 1
 
 #include <stdbool.h>
+#include <stdint.h>
 #include "openvswitch/datapath-protocol.h"
 
 struct ofpbuf;
 
 struct dpif_linux_vport {
-    /* ioctl command argument. */
-    int cmd;
+    /* Generic Netlink header. */
+    uint8_t cmd;
 
     /* odp_vport header. */
     uint32_t dp_idx;
